backend/ml/two_tower_v2/requirements.txt
numpy>=1.26
pandas>=2.2
pyyaml>=6.0
scikit-learn>=1.5
torch>=2.3
onnx>=1.15
onnxruntime>=1.18
pyarrow>=17.0
rich>=13.7
